+++ src/main/java/net/silentchaos512/loot/lib/IBagType.java
@@ -14,6 +14,8 @@ public interface IBagType {
 
     int getBagOverlayColor();
 
+    int getBagStringColor();
+
     String getCustomName();
 
     ResourceLocation getLootTable();